requirements.txt
fsspec
ftfy
pillow
numpy
opencv-python
regex
tqdm
open_clip_torch==2.32.0
safetensors==0.4.3
thop==0.1.1.post2209072238
transformers==4.50.0
timm==1.0.15